Create README.md
I am learning programming to create a free exchange for apps created by users with open source code. The apps can be downloaded by non-creators with a subscription.
The second point of this project is to create apps to teach programming to the people. Making we-app the  modern Encyclopedia of the computer technology.